Added new feature
Concatenation script is now part of of the main program. Accessible in the main screen on  "File" -> "join results".
After analyzing the batch, select "join results" and select the folder named "Results", a "ProSeedling Data.csv" file will be created in the same folder, with all the information gathered from the ".csv" files in the directory.